Use context instead of db call
